+++ utils.c
@@ -395,10 +395,10 @@ value_bytes(Var v)
 }
 
 const char *
-raw_bytes_to_moobinary(const char *buffer, int buflen)
+raw_bytes_to_moobinary(const char *buffer, size_t buflen)
 {
     static Stream *s = 0;
-    int i;
+    size_t i;
 
     if (!s)
 	s = new_stream(100);
@@ -416,7 +416,7 @@ raw_bytes_to_moobinary(const char *buffer, int buflen)
 }
 
 const char *
-moobinary_to_raw_bytes(const char *binary, int *buflen)
+moobinary_to_raw_bytes(const char *binary, size_t *buflen)
 {
     static Stream *s = 0;
     const char *ptr = binary;
@@ -456,7 +456,7 @@ moobinary_to_raw_bytes(const char *binary, int *buflen)
 }
 
 const char *
-recode_chars(const char *chars, int *length,
+recode_chars(const char *chars, size_t *length,
 	     const char *fromcode, const char *tocode)
 {
     iconv_t cd;
